+++ app/controllers/api/pdp_controller.rb
@@ -1,11 +1,15 @@
 # frozen_string_literal: true
 module Api
   class PdpController < Api::ApplicationController
+    before_action :setup_service_and_service_uri
+
     def index
       head(permit? ? :ok : :forbidden)
     end
 
     private
+
+    attr_reader :service, :service_uri
 
     # policies of all matching resources which include the access method in question
     # must allow for access and at least one policy has to exist
@@ -19,19 +23,27 @@ module Api
       every_resource_permitted?(resources)
     end
 
+    def setup_service_and_service_uri
+      return if uri.blank?
+
+      service_and_path = find_service_and_path
+      @service = service_and_path[:service]
+      @service_uri = service_and_path[:service_uri]
+    end
+
+    def find_service_and_path
+      Service.find_each do |service|
+        service_uri = ([service.uri] + service.uri_aliases).
+                      find { |u| uri.starts_with?(u) }
+
+        return { service: service, service_uri: service_uri } if service_uri
+      end
+      {}
+    end
+
     def path
       postfix = uri
-      postfix[(service.uri.length)..-1]
-    end
-
-    def service
-      @service ||= find_service
-    end
-
-    def find_service
-      Service.find_each do |service|
-        return service if ([service.uri] + service.uri_aliases).any? { |u| uri.starts_with?(u) }
-      end
+      postfix[(@service_uri.length)..-1]
     end
 
     def every_resource_permitted?(resources)